[refactor] unified swap related files, updated debug script
